refactor: move HandshakeProcessor and HandshakeAction to utils/, merge into one file
